Remove code duplication in shinyfeedback.js
Introduced a baseInputFeedback object that serves as a base class
for the other xxxInputFeedback implementations. Each impl need only
provide the methods that differ from the base class.

With these changes, I think the Shiny "feedback" custom message
handler can be slightly simplified--it should no longer be necessary
to hide() then show() for cases where feedback is shown twice in a
row for an input.